+++ server/src/main/java/org/realityforge/replicant/server/transport/ReplicantSessionManagerImpl.java
@@ -1,6 +1,5 @@
 package org.realityforge.replicant.server.transport;
 
-import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
@@ -24,8 +23,8 @@ import org.realityforge.replicant.server.ChannelAddress;
 import org.realityforge.replicant.server.ChannelLink;
 import org.realityforge.replicant.server.EntityMessage;
 import org.realityforge.replicant.server.EntityMessageEndpoint;
+import org.realityforge.replicant.server.ServerConstants;
 import org.realityforge.replicant.server.json.JsonEncoder;
-import org.realityforge.replicant.shared.SharedConstants;
 
 /**
  * Base class for session managers.
@@ -241,8 +240,8 @@ public abstract class ReplicantSessionManagerImpl
                                @Nullable final String etag,
                                @Nonnull final ChangeSet changeSet )
   {
-    final String requestID = (String) getRegistry().getResource( SharedConstants.REQUEST_ID_KEY );
-    getRegistry().putResource( SharedConstants.REQUEST_COMPLETE_KEY, Boolean.FALSE );
+    final String requestID = (String) getRegistry().getResource( ServerConstants.REQUEST_ID_KEY );
+    getRegistry().putResource( ServerConstants.REQUEST_COMPLETE_KEY, Boolean.FALSE );
     return session.getQueue().addPacket( requestID, etag, changeSet );
   }
 
@@ -997,6 +996,6 @@ public abstract class ReplicantSessionManagerImpl
   private void setupRegistryContext( @Nonnull final String sessionID )
   {
     //Force the sessionID to the desired session in case call has not been set up by boundary
-    getRegistry().putResource( SharedConstants.SESSION_ID_KEY, sessionID );
+    getRegistry().putResource( ServerConstants.SESSION_ID_KEY, sessionID );
   }
 }