+++ shared/src/main/java/org/realityforge/replicant/shared/SharedConstants.java
@@ -8,24 +8,9 @@ public final class SharedConstants
    */
   public static final String REPLICATION_INVOCATION_KEY = "ReplicationActive";
   /**
-   * Key used to retrieve an opaque identifier for the session from the ReplicantContextHolder.
-   * Used to pass data from the servlet to the EJB.
-   */
-  public static final String SESSION_ID_KEY = "SessionID";
-  /**
-   * Key used to retrieve an opaque identifier for the request from the ReplicantContextHolder.
-   * Used to pass data from the servlet to the EJB.
-   */
-  public static final String REQUEST_ID_KEY = "RequestID";
-  /**
-   * Key used to retrieve a flag whether the request produced a changeset relevant for the initiating session..
-   * Used to pass data from the EJB to the servlet.
-   */
-  public static final String REQUEST_COMPLETE_KEY = "RequestComplete";
-  /**
    * HTTP request header to indicate the session id.
    */
-  public static final String SESSION_ID_HEADER = "X-Replicant-SessionID";
+  public static final String CONNECTION_ID_HEADER = "X-Replicant-SessionID";
   /**
    * HTTP request header to indicate the request id.
    */